Move models related browse components to separate subdir
