+++ frontend/src/metabase/browse/index.tsx
@@ -1,5 +1,5 @@
 export { BrowseMetrics } from "./metrics";
-export { BrowseModels } from "./components/BrowseModels";
+export { BrowseModels } from "./models";
 export { BrowseDatabases } from "./components/BrowseDatabases";
 export { BrowseTables } from "./components/BrowseTables";
 export { default as BrowseSchemas } from "./components/BrowseSchemas";
