+++ frontend/src/metabase/browse/models/index.tsx
@@ -1,0 +1,1 @@
+export { BrowseModels } from "./BrowseModels";
